Move debuging views back to RWGame
